bot/requirements.txt
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
botbuilder-integration-aiohttp>=4.14.0
python-dotenv>=1.0.0
openai>=1.2.0
tiktoken>=0.5.0
